acrn-hypervisor: always run install for hypervisor
do_install should always install hypervisor regardless of
ACRN_FIRMWARE.

acrn.32.out file has elf32-i386 file format in 64-bit build,
which throws Unknown file format error.
And Stripping is already done by source while building. So
'arch' and 'already-stripped' QA checks can be skipped.

remove patch for strip as no longer needed.

also add acrn.32.out to DEPLOY_DIR.

+++ recipes-core/acrn/acrn-hypervisor.bb
@@ -32,15 +32,16 @@ do_compile() {
 }
 
 do_install() {
+	oe_runmake -C hypervisor install install-debug
 	if [ "${ACRN_FIRMWARE}" = "uefi" ]; then
 		oe_runmake -C misc/efi-stub install install-debug
-	else
-		oe_runmake -C hypervisor install install-debug
 	fi
 
 	# Remove sample files
 	rm -rf ${D}${datadir}/acrn
-	rmdir --ignore-fail-on-non-empty ${D}${datadir}
+	if [ "${ACRN_FIRMWARE}" = "uefi" ]; then
+		rmdir --ignore-fail-on-non-empty ${D}${datadir}
+	fi
 }
 
 FILES_${PN} += "${libdir}/acrn/"
@@ -48,9 +49,17 @@ FILES_${PN}-dbg += "${libdir}/acrn/*.efi.*"
 
 addtask deploy after do_install before do_build
 do_deploy() {
+	install -m 0755 ${D}${libdir}/acrn/acrn.${ACRN_BOARD}.${ACRN_FIRMWARE}.${ACRN_SCENARIO}.32.out ${DEPLOYDIR}
+	rm -f ${DEPLOYDIR}/acrn.32.out
+	lnr ${DEPLOYDIR}/acrn.${ACRN_BOARD}.${ACRN_FIRMWARE}.${ACRN_SCENARIO}.32.out ${DEPLOYDIR}/acrn.32.out
 	if [ "${ACRN_FIRMWARE}" = "uefi" ]; then
 		install -m 0755 ${D}${libdir}/acrn/acrn.${ACRN_BOARD}.${ACRN_SCENARIO}.efi ${DEPLOYDIR}
 		rm -f ${DEPLOYDIR}/acrn.efi
 		lnr ${DEPLOYDIR}/acrn.${ACRN_BOARD}.${ACRN_SCENARIO}.efi ${DEPLOYDIR}/acrn.efi
 	fi
 }
+
+# acrn.32.out file has elf32-i386 file format in 64-bit build, which throws Unknown file format error.
+# Stripping is already done by source while building.
+# So 'arch' and 'already-stripped' QA checks can be skipped.
+INSANE_SKIP_${PN} += "arch already-stripped"
